# .migration-backup/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Enterprise Recruitment Automation Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /candidates:
    get:
      operationId: listCandidates
      tags: [candidates]
      summary: List candidates
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: stage
          in: query
          schema:
            type: string
        - name: jobId
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of candidates
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CandidateList"
    post:
      operationId: createCandidate
      tags: [candidates]
      summary: Add a candidate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CandidateInput"
      responses:
        "201":
          description: Created candidate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
  /candidates/{id}:
    get:
      operationId: getCandidate
      tags: [candidates]
      summary: Get candidate by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Candidate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
    patch:
      operationId: updateCandidate
      tags: [candidates]
      summary: Update a candidate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CandidateUpdate"
      responses:
        "200":
          description: Updated candidate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
    delete:
      operationId: deleteCandidate
      tags: [candidates]
      summary: Delete candidate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /candidates/{id}/ai-analysis:
    get:
      operationId: getCandidateAiAnalysis
      tags: [candidates]
      summary: Get AI resume analysis for a candidate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: AI analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiAnalysis"
    post:
      operationId: generateCandidateAiAnalysis
      tags: [candidates]
      summary: Trigger AI resume analysis for a candidate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: AI analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiAnalysis"

  # ─── APPLICATIONS ─────────────────────────────────────────────────────────────
...
  /interviews:
    get:
      operationId: listInterviews
      tags: [interviews]
      summary: List interviews
      parameters:
        - name: applicationId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of interviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Interview"
    post:
      operationId: createInterview
      tags: [interviews]
      summary: Schedule an interview
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InterviewInput"
      responses:
        "201":
          description: Created interview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Interview"
  /interviews/{id}:
    get:
      operationId: getInterview
      tags: [interviews]
      summary: Get interview by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Interview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Interview"
    patch:
      operationId: updateInterview
      tags: [interviews]
      summary: Update an interview
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InterviewUpdate"
      responses:
        "200":
          description: Updated interview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Interview"
  /interviews/{id}/feedback:
    get:
      operationId: getInterviewFeedback
      tags: [interviews]
      summary: Get feedback for an interview
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Feedback list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InterviewFeedback"
    post:
      operationId: submitInterviewFeedback
      tags: [interviews]
      summary: Submit interview feedback
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InterviewFeedbackInput"
      responses:
        "201":
          description: Created feedback
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InterviewFeedback"
  /interviews/{id}/ai-summary:
    get:
      operationId: getInterviewAiSummary
      tags: [interviews]
      summary: Get AI-generated interview summary
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: AI summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiInterviewSummary"

  # ─── OFFERS ──────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Job:
      type: object
      required: [id, title, department, type, status, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        department:
          type: string
        description:
          type: ["string", "null"]
        competencies:
          type: array
          items:
            type: string
        type:
          type: string
          enum: [internal, external, both]
        status:
          type: string
          enum: [draft, active, closed, expired]
        source:
          type: ["string", "null"]
        expiryDate:
          type: ["string", "null"]
        requisitionId:
          type: ["integer", "null"]
        applicationCount:
          type: integer
        createdAt:
          type: string
...
    Candidate:
      type: object
      required: [id, firstName, lastName, email, source, createdAt]
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        currentRole:
          type: ["string", "null"]
        currentCompany:
          type: ["string", "null"]
        yearsOfExperience:
          type: ["number", "null"]
        skills:
          type: array
          items:
            type: string
        source:
          type: string
        resumeUrl:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        aiMatchScore:
          type: ["number", "null"]
        createdAt:
          type: string
    CandidateInput:
      type: object
      required: [firstName, lastName, email, source]
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        currentRole:
          type: ["string", "null"]
        currentCompany:
          type: ["string", "null"]
        yearsOfExperience:
          type: ["number", "null"]
        skills:
          type: array
          items:
            type: string
        source:
          type: string
        resumeUrl:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    CandidateUpdate:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: ["string", "null"]
        currentRole:
          type: ["string", "null"]
        currentCompany:
          type: ["string", "null"]
        yearsOfExperience:
          type: ["number", "null"]
        skills:
          type: array
          items:
            type: string
        notes:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
    CandidateList:
      type: object
      required: [data, total]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Candidate"
        total:
          type: integer
    AiAnalysis:
      type: object
      required: [candidateId, summary, matchScore, strengths, gaps, recommendation, generatedAt]
      properties:
        candidateId:
          type: integer
        summary:
          type: string
        matchScore:
          type: number
        extractedSkills:
          type: array
          items:
            type: string
        strengths:
          type: array
          items:
            type: string
        gaps:
          type: array
          items:
            type: string
        recommendation:
          type: string
          enum: [strong_hire, hire, maybe, no_hire]
        recommendationRationale:
          type: string
        generatedAt:
          type: string

    # APPLICATIONS
    Application:
      type: object
      required: [id, candidateId, jobId, stage, appliedAt]
      properties:
        id:
          type: integer
        candidateId:
          type: integer
        jobId:
          type: integer
        candidate:
          $ref: "#/components/schemas/Candidate"
        job:
          $ref: "#/components/schemas/Job"
        stage:
          type: string
          enum: [applied, under_review, longlisted, shortlisted, interview_scheduled, interviewed, offer_pending, offered, hired, rejected]
        notes:
          type: ["string", "null"]
        rejectionReason:
          type: ["string", "null"]
        appliedAt:
          type: string
        updatedAt:
          type: string
...
    Interview:
      type: object
      required: [id, applicationId, scheduledAt, type, status, createdAt]
      properties:
        id:
          type: integer
        applicationId:
          type: integer
        application:
          $ref: "#/components/schemas/Application"
        scheduledAt:
          type: string
        duration:
          type: integer
        type:
          type: string
          enum: [phone_screen, technical, behavioral, panel, final]
        status:
          type: string
          enum: [scheduled, completed, cancelled, rescheduled]
        location:
          type: ["string", "null"]
        meetingUrl:
          type: ["string", "null"]
        panelMembers:
          type: array
          items:
            type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    InterviewInput:
      type: object
      required: [applicationId, scheduledAt, type]
      properties:
        applicationId:
          type: integer
        scheduledAt:
          type: string
        duration:
          type: integer
        type:
          type: string
        location:
          type: ["string", "null"]
        meetingUrl:
          type: ["string", "null"]
        panelMembers:
          type: array
          items:
            type: string
        notes:
          type: ["string", "null"]
    InterviewUpdate:
      type: object
      properties:
        scheduledAt:
          type: string
        status:
          type: string
        location:
          type: ["string", "null"]
        meetingUrl:
          type: ["string", "null"]
        panelMembers:
          type: array
          items:
            type: string
        notes:
          type: ["string", "null"]
    InterviewFeedback:
      type: object
      required: [id, interviewId, interviewerName, recommendation, overallScore, createdAt]
      properties:
        id:
          type: integer
        interviewId:
          type: integer
        interviewerName:
          type: string
        technicalScore:
          type: ["number", "null"]
        behavioralScore:
          type: ["number", "null"]
        overallScore:
          type: number
        strengths:
          type: ["string", "null"]
        concerns:
          type: ["string", "null"]
        recommendation:
          type: string
          enum: [strong_hire, hire, maybe, no_hire]
        comments:
          type: ["string", "null"]
        createdAt:
          type: string
    InterviewFeedbackInput:
      type: object
      required: [interviewerName, recommendation, overallScore]
      properties:
        interviewerName:
          type: string
        technicalScore:
          type: ["number", "null"]
        behavioralScore:
          type: ["number", "null"]
        overallScore:
          type: number
        strengths:
          type: ["string", "null"]
        concerns:
          type: ["string", "null"]
        recommendation:
          type: string
        comments:
          type: ["string", "null"]
    AiInterviewSummary:
      type: object
      required: [interviewId, summary, overallAssessment, generatedAt]
      properties:
        interviewId:
          type: integer
        summary:
          type: string
        keyStrengths:
          type: array
          items:
            type: string
        keyGaps:
          type: array
          items:
            type: string
        overallAssessment:
          type: string
        panelConsensus:
          type: string
        generatedAt:
          type: string

    # OFFERS
...