# .migration-backup/lib/api-zod/src/generated/types/candidateList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Enterprise Recruitment Automation Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Candidate } from "./candidate";

export interface CandidateList {
  data: Candidate[];
  total: number;
}