# .migration-backup/lib/api-zod/src/generated/types/interviewFeedback.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Enterprise Recruitment Automation Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { InterviewFeedbackRecommendation } from "./interviewFeedbackRecommendation";

export interface InterviewFeedback {
  id: number;
  interviewId: number;
  interviewerName: string;
  /** @nullable */
  technicalScore?: number | null;
  /** @nullable */
  behavioralScore?: number | null;
  overallScore: number;
  /** @nullable */
  strengths?: string | null;
  /** @nullable */
  concerns?: string | null;
  recommendation: InterviewFeedbackRecommendation;
  /** @nullable */
  comments?: string | null;
  createdAt: string;
}