# .migration-backup/lib/api-zod/src/generated/types/interviewInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Enterprise Recruitment Automation Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface InterviewInput {
  applicationId: number;
  scheduledAt: string;
  duration?: number;
  type: string;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  meetingUrl?: string | null;
  panelMembers?: string[];
  /** @nullable */
  notes?: string | null;
}